width bug fix

+++ srcs/printf/format2.c
@@ -6,11 +6,12 @@
 /*   By: orenkay <orenkay@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/13 17:46:10 by spopieul          #+#    #+#             */
-/*   Updated: 2018/02/25 18:01:29 by orenkay          ###   ########.fr       */
+/*   Updated: 2018/02/26 21:44:51 by orenkay          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./includes/ft_printf.h"
+
 
 void		ft_pf_format_s(t_pf_state *state)
 {
@@ -48,7 +49,7 @@ void		ft_pf_format_signed(t_pf_state *state)
 	if (*data.value == '0' && state->precision == 0)
 		data.len = 0;
 	data.bpad = "";
-	data.precision = FT_MIN(state->precision - data.len, 0);
+	data.precision = FT_MIN(state->precision - (int)data.len, 0);
 	data.width = ft_pf_get_pad_width(state, &data);
 	data.pchar = (FT_MASK_EQ(state->flags, M_FLAG_ZERO) ? "0" : " ");
 	if (state->precision > -1)
@@ -69,7 +70,7 @@ void		ft_pf_format_unsigned(t_pf_state *state)
 	if (*data.value == '0' && state->precision == 0)
 		data.len = 0;
 	data.bpad = ft_pf_get_base_padding(state, &data);
-	data.precision = FT_MIN(state->precision - data.len, 0);
+	data.precision = FT_MIN(state->precision - (int)data.len, 0);
 	data.width = ft_pf_get_pad_width(state, &data);
 	if (ft_tolower(state->specifier) == 'o')
 		data.precision -= (int)ft_strlen(data.bpad);
